// src/app/common/enum/notificaciones.enum.ts
export enum ENotificacion {
  FORMATOS_FIRMADOS = 1,
  SOLICITUD_FIRMA_ASEGURADO = 2,
  SOLICITUD_FIRMA_EJECUTIVO = 3,
  SOLICITUD = 4,
  CONFIRMACION_ENTREGA = 5,
  REVISION_DOCUMENTAL = 6,
  DATOS_CONTACTO = 7,
  NO_CONTINUA_PROCESO = 8,
  RECEPCION_DOCUMENTOS = 9,
  RECOLECCION_FISICOS = 10,
  FISICOS_ENVIADOS = 11,
  REVISION_FORMATOS = 12,
  FIRMA_DOCUMENTOS = 13,
  REVISION_DOCUMENTACION_FISICA = 14,
  GENERACION_DE_FORMATOS = 15
}


export enum EBitNotificacionType {
  MARSH_NOTIFICACION_ACTIVAR_CUENTA = 'MARSH_NOTIFICACION_ACTIVAR_CUENTA',
  MARSH_NOTIFICACION_CODIGO_MFA = 'MARSH_NOTIFICACION_CODIGO_MFA',
  MARSH_NOTIFICACION_RECUPERAR_PASSWORD = 'MARSH_NOTIFICACION_RECUPERAR_PASSWORD',
  MARSH_NOTIFICACION_ESTATUS_CARGA_MASIVA = 'MARSH_NOTIFICACION_ESTATUS_CARGA_MASIVA',
  MARSH_NOTIFICACION_SOLICITUD_FIRMA_ASEGURADO = 'MARSH_NOTIFICACION_SOLICITUD_FIRMA_ASEGURADO',
  MARSH_NOTIFICACION_DATOS_CONTACTO = 'MARSH_NOTIFICACION_DATOS_CONTACTO',
  MARSH_NOTIFICACION_NO_CONTINUA_PROCESO = 'MARSH_NOTIFICACION_NO_CONTINUA_PROCESO',
  MARSH_NOTIFICACION_FORMATOS_FIRMADOS = 'MARSH_NOTIFICACION_FORMATOS_FIRMADOS',
  MARSH_NOTIFICACION_SOLICITUD = 'MARSH_NOTIFICACION_SOLICITUD',
  MARSH_NOTIFICACION_REVISION_DOCUMENTAL = 'MARSH_NOTIFICACION_REVISION_DOCUMENTAL',
  MARSH_NOTIFICACION_RECEPCION_DE_DOCUMENTOS = 'MARSH_NOTIFICACION_RECEPCION_DE_DOCUMENTOS',
  MARSH_NOTIFICACION_SOLICITUD_FIRMA_EJECUTIVO = 'MARSH_NOTIFICACION_SOLICITUD_FIRMA_EJECUTIVO',
  MARSH_NOTIFICACION_RECOLECCION_FISICOS = 'MARSH_NOTIFICACION_RECOLECCION_FISICOS',
  MARSH_NOTIFICACION_REVISION_DE_FORMATOS = 'MARSH_NOTIFICACION_REVISION_DE_FORMATOS',
  MARSH_NOTIFICACION_FIRMA_DE_DOCUMENTOS = 'MARSH_NOTIFICACION_FIRMA_DE_DOCUMENTOS',
  MARSH_NOTIFICACION_FISICOS_ENVIADOS = 'MARSH_NOTIFICACION_FISICOS_ENVIADOS',
  MARSH_NOTIFICACION_REVISION_DOCUMENTACION_FISICA = 'MARSH_NOTIFICACION_REVISION_DOCUMENTACION_FISICA',
  MARSH_NOTIFICACION_GENERACION_FORMATOS = 'MARSH_NOTIFICACION_GENERACION_FORMATOS',
  MARSH_NOTIFICACION_CONFIRMACION_ENTREGA = 'MARSH_NOTIFICACION_CONFIRMACION_ENTREGA',

}